Fix lint and darwin tests
