+++ pkg/tools/otelcol-config/darwin_test.go
@@ -1,0 +1,5 @@
+//go:build darwin
+
+package main
+
+const hostmetricsFilename = hostmetricsDarwin
