+++ pkg/tools/otelcol-config/linux_test.go
@@ -1,0 +1,5 @@
+//go:build linux
+
+package main
+
+const hostmetricsFilename = hostmetricsLinux
