+++ pkg/tools/otelcol-config/main_test.go
@@ -1,10 +1,6 @@
 package main
 
 import (
-	"fmt"
-	"io/fs"
-	"os"
-	"path/filepath"
 	"testing"
 
 	"github.com/google/go-cmp/cmp"
@@ -21,75 +17,5 @@ func TestGetSortedActions(t *testing.T) {
 	exp := []string{flagAddTag, flagSetInstallationToken, flagReadKV}
 	if got, want := sortedActions, exp; !cmp.Equal(got, want) {
 		t.Errorf("actions not correctly sorted: got %v, want %v", got, want)
-	}
-}
-
-func TestHostMetricsLinker(t *testing.T) {
-	dir, err := os.MkdirTemp("", "otelcol-config-test")
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer os.RemoveAll(dir)
-
-	values := &flagValues{
-		ConfigDir: dir,
-	}
-
-	availPath := filepath.Join(dir, ConfDotDAvailable)
-	if err := os.Mkdir(availPath, 0700); err != nil {
-		t.Fatal(err)
-	}
-
-	confdPath := filepath.Join(dir, ConfDotD)
-	if err := os.Mkdir(confdPath, 0700); err != nil {
-	}
-
-	configPath := filepath.Join(availPath, hostmetricsLinux)
-	if err := os.WriteFile(configPath, []byte("configuration: yes"), 0600); err != nil {
-		t.Fatal(err)
-	}
-
-	linker := getHostMetricsLinker(values)
-
-	if err := linker(); err != nil {
-		t.Fatal(fmt.Errorf("can't link config path %s: %s", configPath, err))
-	}
-
-	linkPath := filepath.Join(dir, ConfDotD, hostmetricsLinux)
-
-	stat, err := os.Lstat(linkPath)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if stat.Mode().Type() != fs.ModeSymlink {
-		t.Error("not a symlink")
-	}
-
-	// we expect that calling the linker again succeeds even if the link exists
-	if err := linker(); err != nil {
-		t.Fatal(fmt.Errorf("can't link config path %s: %s", configPath, err))
-	}
-
-	unlinker := getHostMetricsUnlinker(values)
-
-	// unlink the link
-	if err := unlinker(); err != nil {
-		t.Fatal(err)
-	}
-
-	// verify the link is gone
-	_, err = os.Lstat(linkPath)
-	if err == nil {
-		t.Fatal("expected non-nil error")
-	}
-	if _, ok := err.(*os.PathError); !ok {
-		t.Fatal(err)
-	}
-
-	// verify the linked file is still there
-	_, err = os.Stat(configPath)
-	if err != nil {
-		t.Fatal(err)
 	}
 }
